refactor: date의 확장을 고려해 폴더구조를 변경
