+++ src/date/days.constants.ts
@@ -1,0 +1,19 @@
+export const DAYS_MAP = {
+  1: '하루',
+  2: '이틀',
+  3: '사흘',
+  4: '나흘',
+  5: '닷새',
+  6: '엿새',
+  7: '이레',
+  8: '여드레',
+  9: '아흐레',
+  10: '열',
+  20: '스무',
+} as const;
+
+export const DAYS_ONLY_TENS_MAP = {
+  10: '열흘',
+  20: '스무날',
+  30: '서른날',
+} as const;
